[REF] Move transport request into joint_buying_base module
