+++ joint_buying_base/__manifest__.py
@@ -11,6 +11,8 @@
     "depends": [
         "base",
         "mail",
+        "web",
+        "decimal_precision",
         # OCA
         "base_geolocalize_company",
         "res_company_code",
@@ -21,7 +23,9 @@
         # GRAP
         "name_search_reset_res_partner",
     ],
-    "external_dependencies": {"python": ["openupgradelib", "geopy", "bokeh", "pandas"]},
+    "external_dependencies": {
+        "python": ["openupgradelib", "geopy", "bokeh", "pandas", "treelib"]
+    },
     "data": [
         "security/ir_module_category.xml",
         "security/res_groups.xml",
@@ -38,7 +42,12 @@
         "views/view_joint_buying_tour_type.xml",
         "views/view_joint_buying_tour.xml",
         "views/view_joint_buying_tour_line.xml",
+        "views/view_joint_buying_transport_request.xml",
+        "views/view_joint_buying_transport_request_line.xml",
         "wizards/view_joint_buying_wizard_set_tour.xml",
+        "wizards/joint_buying_wizard_find_route.xml",
+        "reports/report_joint_buying_tour.xml",
+        "reports/report.xml",
         "views/templates.xml",
     ],
     "demo": [
@@ -49,6 +58,7 @@
         "demo/joint_buying_carrier.xml",
         "demo/joint_buying_tour_type.xml",
         "demo/joint_buying_tour.xml",
+        "demo/joint_buying_transport_request.xml",
     ],
     "post_init_hook": "_create_joint_buying_partner_for_companies",
     "installable": True,
