+++ joint_buying_base/models/joint_buying_transport_request_line.py
@@ -4,9 +4,7 @@
 
 from odoo import fields, models
 
-from odoo.addons.joint_buying_base.models.res_partner import (
-    _JOINT_BUYING_PARTNER_CONTEXT,
-)
+from .res_partner import _JOINT_BUYING_PARTNER_CONTEXT
 
 
 class JointBuyingTransportRequestLine(models.Model):
