+++ joint_buying_base/security/ir.model.access.csv
@@ -6,6 +6,8 @@ access_joint_buying_tour_user,access_joint_buying_tour_user,model_joint_buying_t
 access_joint_buying_tour_manager,access_joint_buying_tour_manager,model_joint_buying_tour,joint_buying_base.group_joint_buying_manager,1,1,1,1
 access_joint_buying_tour_line_user,access_joint_buying_tour_line_user,model_joint_buying_tour_line,joint_buying_base.group_joint_buying_user,1,,,
 access_joint_buying_tour_line_manager,access_joint_buying_tour_line_manager,model_joint_buying_tour_line,joint_buying_base.group_joint_buying_manager,1,1,1,1
+access_joint_buying_transport_request_user,access_joint_buying_transport_request_user,model_joint_buying_transport_request,joint_buying_base.group_joint_buying_user,1,1,1,1
+access_joint_buying_transport_request_line_user,access_joint_buying_transport_request_line_user,model_joint_buying_transport_request_line,joint_buying_base.group_joint_buying_user,1,1,1,1
 access_joint_buying_carrier_user,access_joint_buying_carrier_user,model_joint_buying_carrier,joint_buying_base.group_joint_buying_user,1,,,
 access_joint_buying_carrier_manager,access_joint_buying_carrier_manager,model_joint_buying_carrier,joint_buying_base.group_joint_buying_manager,1,1,1,1
 access_joint_buying_tour_type_user,access_joint_buying_tour_type_user,model_joint_buying_tour_type,joint_buying_base.group_joint_buying_user,1,,,
