+++ joint_buying_base/tests/__init__.py
@@ -4,4 +4,5 @@ from . import test_joint_buying_check_access_mixin
 from . import test_joint_buying_mixin
 from . import test_partner_global_local
 from . import test_partner_subscription
+from . import test_joint_buying_wizard_find_route
 from . import test_tour
