+++ joint_buying_base/wizards/__init__.py
@@ -1,2 +1,3 @@
+from . import joint_buying_wizard_find_route
 from . import joint_buying_wizard_set_tour
 from . import joint_buying_wizard_set_tour_line
