+++ joint_buying_base/wizards/joint_buying_wizard_find_route.py
@@ -9,9 +9,7 @@ from treelib import Tree
 
 from odoo import api, fields, models
 
-from odoo.addons.joint_buying_base.models.res_partner import (
-    _JOINT_BUYING_PARTNER_CONTEXT,
-)
+from ..models.res_partner import _JOINT_BUYING_PARTNER_CONTEXT
 
 
 class JointBuyingWizardFindRoute(models.TransientModel):
