+++ joint_buying_product/__init__.py
@@ -1,3 +1,4 @@
 from . import models
 from . import wizards
 from .pre_init_hook import pre_init_product_db
+from .uninstall_hook import uninstall_product_db
