+++ joint_buying_product/__manifest__.py
@@ -20,12 +20,11 @@
         "product_uom_package",
     ],
     "pre_init_hook": "pre_init_product_db",
-    "external_dependencies": {"python": ["treelib"]},
+    "uninstall_hook": "uninstall_product_db",
     "data": [
         "security/ir.model.access.csv",
         "views/menu.xml",
         "wizards/joint_buying_wizard_create_order.xml",
-        "wizards/joint_buying_wizard_find_route.xml",
         "wizards/joint_buying_wizard_update_order_grouped.xml",
         "wizards/mail_compose_message_purchase_order_grouped.xml",
         "views/view_res_config_settings.xml",
@@ -37,11 +36,8 @@
         "views/view_joint_buying_purchase_order_grouped.xml",
         "views/view_joint_buying_purchase_order.xml",
         "views/view_joint_buying_tour.xml",
-        "views/view_joint_buying_transport_request.xml",
-        "views/view_joint_buying_transport_request_line.xml",
         "views/view_res_partner.xml",
         "reports/report_joint_buying_purchase_order_grouped.xml",
-        "reports/report_joint_buying_tour.xml",
         "reports/report.xml",
         "data/ir_cron.xml",
         "data/ir_config_parameter.xml",
@@ -57,7 +53,6 @@
         "demo/product_product.xml",
         "demo/res_partner.xml",
         "demo/joint_buying_frequency.xml",
-        "demo/joint_buying_transport_request.xml",
         "demo/joint_buying_purchase_order_grouped.xml",
         "demo/ir_cron.xml",
     ],
