+++ joint_buying_product/models/__init__.py
@@ -10,7 +10,5 @@ from . import joint_buying_purchase_order_grouped
 from . import joint_buying_purchase_order_grouped_line
 from . import joint_buying_purchase_order
 from . import joint_buying_purchase_order_line
-from . import joint_buying_tour_line
 from . import joint_buying_transport_request
-from . import joint_buying_transport_request_line
 from . import res_config_settings
