+++ joint_buying_product/reports/report.xml
@@ -10,14 +10,4 @@
         print_report_name="('%s - %s - Grouped Order (%s)' % (object.name, object.supplier_id.name, object.deposit_date)).replace('/', '_')"
     />
 
-    <report
-        id="action_report_joint_buying_tour"
-        string="Tour"
-        model="joint.buying.tour"
-        report_type="qweb-pdf"
-        file="joint_buying_product.report_joint_buying_tour"
-        name="joint_buying_product.report_joint_buying_tour"
-        print_report_name="('Tour %s' % (object.name)).replace('/', '_')"
-    />
-
 </odoo>
