+++ joint_buying_product/security/ir.model.access.csv
@@ -4,8 +4,6 @@ access_joint_buying_purchase_order,access_joint_buying_purchase_order,model_join
 access_joint_buying_purchase_order_grouped,access_joint_buying_purchase_order_grouped,model_joint_buying_purchase_order_grouped,joint_buying_base.group_joint_buying_user,1,1,1,
 access_joint_buying_purchase_order_grouped_manager,access_joint_buying_purchase_order_grouped_manager,model_joint_buying_purchase_order_grouped,joint_buying_base.group_joint_buying_manager,1,1,1,1
 access_joint_buying_frequency_user,access_joint_buying_frequency_user,model_joint_buying_frequency,joint_buying_base.group_joint_buying_user,1,1,1,1
-access_joint_buying_transport_request_user,access_joint_buying_transport_request_user,model_joint_buying_transport_request,joint_buying_base.group_joint_buying_user,1,1,1,1
-access_joint_buying_transport_request_line_user,access_joint_buying_transport_request_line_user,model_joint_buying_transport_request_line,joint_buying_base.group_joint_buying_user,1,1,1,1
 access_joint_buying_category_user,access_joint_buying_category_user,model_joint_buying_category,joint_buying_base.group_joint_buying_user,1,1,1,1
 access_product_product,access_product_product,product.model_product_product,joint_buying_base.group_joint_buying_user,1,1,1,1
 access_product_template,access_product_template,product.model_product_template,joint_buying_base.group_joint_buying_user,1,1,1,1
