+++ joint_buying_product/tests/__init__.py
@@ -1,7 +1,6 @@
 from . import test_abstract
 from . import test_joint_buying_purchase_order
 from . import test_joint_buying_transport_request
-from . import test_joint_buying_wizard_find_route
 from . import test_product
 from . import test_check_access_product
 from . import test_check_access_joint_buying_purchase
