+++ joint_buying_product/wizards/__init__.py
@@ -1,6 +1,5 @@
 from . import joint_buying_wizard_create_order
 from . import joint_buying_wizard_create_order_line
-from . import joint_buying_wizard_find_route
 from . import joint_buying_wizard_update_order_grouped
 
 from . import mail_compose_message_purchase_order_grouped
