+++ joint_buying_sale/views/view_joint_buying_transport_request.xml
@@ -8,7 +8,7 @@ License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
     <record id="view_joint_buying_transport_request_form" model="ir.ui.view">
         <field name="model">joint.buying.transport.request</field>
-        <field name="inherit_id" ref="joint_buying_product.view_joint_buying_transport_request_form"/>
+        <field name="inherit_id" ref="joint_buying_base.view_joint_buying_transport_request_form"/>
         <field name="arch" type="xml">
             <field name="order_id" position="after">
                 <field name="sale_order_id" attrs="{'invisible': [('sale_order_id', '=', False)]}"/>
